mongooseim: Make config file a giant property list
So far we created the configuration file using simple string
concatenation and/or antiquotations.

However, if we want to avoid continuously shoveling more and more values
into strings this is going to get _very_ messy at the end and it's also
even more difficult to change because we need to verify whether the
generated configuration values are correct.

Especially when we have host-specific configuration values, using string
concatenation is going to be hell on earth, because the string
comprising the config file so far is a list of Erlang terms separated by
dots.

If we want to have host-specific configuration values, we need to have
almost the same but separated by commas and surrounded by square
brackets.

The new configuration values for extraConfig currently are especially
ugly and very uneasy to read, so we need to convert them to NixOS
options very soon.

Also the configuration values for listeners and modules are currently
raw values, so we need to refactor them as well.

Nevertheless, this is a step forward and we need to run our tests first
and fix stuff before starting to restructure other stuff so that we can
make sure we don't introduce regressions we can't track anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>