services/mongooseim: Add a new shapers option
The shapers that we were defining using extraConfig had been defined in
the wrong way, so we got something like this for example:

{shaper, fast, [{maxrate, 50000}]}.
{shaper, normal, [{maxrate, 1000}]}.

However, this is not correct and these lines really should be:

{shaper, fast, {maxrate, 50000}}.
{shaper, normal, {maxrate, 1000}}.

But instead of just fixing our extraConfigs, let's create a proper
option instead so that we don't run into such an issue again.

I've also checked the code on whether it's possible to specify other
values than none or {maxrate, N}, but it isn't the case, so let's keep
the option simple.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>