mongooseim: Add NixOS options for ACL patterns
Another thing we got wrong, because the actual {acl, ..., ...}
definition requires the third argument to be a single pattern rather
than being a list of patterns.

For multiple matches in MongooseIMs config we'd need to define multiple
entries for the same ACL.

However right now we don't need this in production and it also doesn't
work well with NixOS 16.09, because it doesn't have the types.either
improvement that NixOS 17.03 will have (or better: already has in NixOS
Unstable).

The actual pattern implementation is done by transforming arbitrary sets
of the options defined in the acl submodule to the values needed by the
real MongooseIM config.

We don't use the same values as the upstream project, because they're
using tuples of arbitrary length for those patterns which do not map
very well to Nix types.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>