mongoosim/acl: Allow to set empty patterns
At the moment this doesn't have any benefit, but we're going to validate
the pattern names very soon, so it's cruicial to have a way to specify
placeholder pattern rules.

These placeholders won't generate any entries in the resulting
configuration file, so they can for example used for specifying ACLs for
banned/blocked users which then can be later filled with the
corresponding values at runtime.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>