mongooseim: Allow to define ACL rules
This generates {access, ..., ...} tuples in the resulting configuration
file.

On our side these access tuples are called acl.rules, because "access"
would imply that it's solely for access control, which isn't the case.

I suspect the reason why the access tuples are called like this is
historic and they were for access control in the first place but shapers
and limits were added later.

Nevertheless, when trying to come up with a good Nix representation for
these options, I went for something like this:

  { foo = [ { allow = false; match = "outsiders"; }
            { allow = true; }
          ];
  }

Which results in an access tuple like this:

  {access, foo, [{deny, outsiders}, {allow, all}]}

Our syntax is a bit more verbose, but we have other benefits, such as we
can validate the types and also check whether the pattern the rule is
referring to is actually defined.

In regards to shoveling those access rules into extraConfig our syntax
is however way less verbose and also way better to read, so it's a
tremendous improvement here.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>