services/mongooseim: Flesh out S2S TLS options
The last two options for domain_certfile and s2s_ciphers are now gone in
extraConfig of our main config in xmpp.nix, which means that the only
thing left we need to flesh out are vhost options.

This implementation is pretty straightforward except that we use a list
of ciphers instead of a plain cipher string already containing colons,
because it makes it easier to do overrides and ordering using the NixOS
module system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>