mongooseim: Remove __raw for listeners/modules
Instead of returning a generatedConfig value which is a plain string we
now return structured information about these modules which in turn can
be used directly in the configuration generator in settings.nix.

So having this large churn of options available in the generator opens a
few new possibilities, like for example automatically pretty-printing
the configuration file or even overriding certain values from the
top-level.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>