services/mongooseim: Allow per-vhost settings
This is a lot more involved and still far from complete, because if we
want to have per-host settings that are going to be appended to the
global options we need to add a way to order Erlang property lists.

Nevertheless, this refactors the whole settings module which now uses an
internal "expression" option to pass through the generated configuration
values and they're merged in the top-level default.nix instead of within
settings.nix.

We also get rid of all the __raw configuration values that were used in
settings.nix, because we need to check whether an option is at its
default and we don't want apply or type merge to interfere with that.

The reason we need to check whether the option is at its default value
is that we want the per-vhost settings to be empty by default.

Another major change which fortunately doesn't affect configuration is
that ./modules.nix is no longer a submodule. It didn't make sense in the
first place to actually be one, especially because it's hard to reason
about the way overriding is done on individual module options.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>