xmpp: Remove limitations on max stanza size
We only allowed a maximum of 65536 bytes on C2S, which sometimes is a
bit very low and we get errors because of that resulting in unexpected
situations because some clients tend to not cope with those errors very
well and immediately disconnect.

For limiting the amount of data, we still have the shaper which I think
should be enough.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>