Add a small module for running Boulder ACME Server
This module is going to be used for tests where we need to impersonate
an ACME service such as the one from Letsencrypt within VM tests, which
is the reason why this module is a bit ugly (I only care if it's working
not if it's beautiful).

While the module isn't used anywhere, it will serve as a pluggable
module for testing whether ACME works properly to fetch certificates and
also as a replacement for our snakeoil certificate generator.

For now we have our vhost module which takes care of setting SSL
certificates and keys and the goal is to have an ACME module that's not
only much simpler than the one we have in <nixpkgs> but it's also tied
into the vhost module so that all you have to do is define another
domain name and generating DNS record, configuring IPs and of course SSL
certificates is done automatically without needing to add another line
of configuration.

But back to the newly added module:

The main point which is missing is abstraction in terms of Nix, so that
the default domains which come with the Boulder test configuration can
be changed via Nix attributes.

As an example, the main hostname for the service is currently "boulder",
se we might want to call it "letsencrypt.org" and inject a snakeoil CA
certificate to the CA bundles of all test nodes. This makes it possible
to test certificate creation without the need to alter the ACME clients
or creating ugly /etc/hosts entries.

Right now the path (tests/ssl/boulder.nix) is only temporary and will be
changed as soon as we have decided on a good namespace for the ACME
tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>