boulder: Refactor to be fully usable in VM tests
We only had a small Boulder test setup running which mostly only
resolves to 127.0.0.1, which is not very useful when we have a network
of several test VMs at our disposal and actually *want* to use such a
network so we can run tests that are as close to production as possible.

Now we have a fully pluggable module which automatically discovers in
which network it's used via the nodes attribute.

The test environment of Boulder used "dns-test-srv", which is a fake DNS
server that's resolving almost everything to 127.0.0.1. On our setup
this is not useful, so instead we're now running a local BIND name
server which has a fake root zone and uses the mentioned node attribute
to automatically discover other zones in the network of machines and
generate delegations from the root zone to the respective zones with the
primaryIPAddress of the node.

Apart from that the only thing that's a bit ugly right now is that we
have to use headcounter.snakeOilCaCert in other nodes in order to use
the snakeoil CA certificate we're using now.

We want to use real letsencrypt.org FQDNs here, so we can't get away
with the snakeoil test certificates from the upstream project but now
roll our own.

This not only has the benefit that we can easily pass the snakeoil
certificate to other nodes, but we can (and do) also use it for an nginx
proxy that's now serving HTTPS for the Boulder web front end.

I've tested this with a small rudimentary simp_le test case and
certificate issuance now works like a charm :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>