services/dyndns: Retry binding to slave sockets
If there is an error on binding on the slave-facing sockets on the
master, the thread which serves the slaves would simply exit without
notice.

In production this happens either if there's an error bringing up the
tunnels or if the tunnel simply isn't online yet.

Now we're running the listener again after three seconds if an exception
was thrown while trying to bind the socket.

I've also added a few regression tests, which simulate this by creating
a listening socket without actually running accept, so that the socket
is blocked and the master should get back to work as soon as we remove
the blocker.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>