modules/webspace: Drop SSH1 stuff in openssh.patch
SSH version 1 has long been deprecated in OpenSSH and now the portion of
patch which does the same in the v1 authentication code which we do in
the v2 is no longer there.

See the release notes for deprecation:

https://www.openssh.com/txt/release-7.4

We really didn't use the first hunk of that patch anyway, so we can
safely drop it without shedding a tear.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>