common-machines: Don't use fetchTarball
We are going to use this for tests across the whole deployment and we're
running this on our Hydra. However, Hydra doesn't allow fetchTarball, so
we use fetchFromGitHub instead.

This also has the advantage that Nix doesn't need to check the tarball
again and again, because we have a fixed sha256 hash here.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>