hydra: Update to latest upstream master
Had this already updated in production, so for the next incarnation of
the Headcounter deployment we definitly want to have either the same
version or a newer one.

I'm heading for a newer one, because the same version needed to be
patched up a whole lot and it *seems* that most of these fixes already
went into master.

Another thing that has changed is that we now have to use hydra-dev as
the service name, because Hydra is now included in <nixpkgs>.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>