modules/testing: Add a module for NixOps testing
NixOps has a few more options for machines and resources within its
deployment than the NixOS VM tests have, so this module declares these
options and implements some of these options for VM tests while ignoring
others.

One of the options that are now available is encrypted SSH tunnels
specified with deployment.encryptedLinksTo. I'm mostly replicating what
NixOps does when creating the physical machine specificaton here with
the difference that we use snakeoil SSH keypairs, so they end up being
in the store.

And the second option that we now have - which is kinda related - is the
resources.sshKeyPairs option that's *outside* of the machine
specification and is passed through module arguments to the network.

So the latter is the only thing that makes this module a bit less
pluggable, because we need to reinject the resources into all of the
machines.

Apart from that however, everything else is taken care automatically.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>