tests/headcounter: Include all machines in tests
This causes makeHeadcounterTest to test the *full* network instead of
just a small portion of ultron (the XMPP config).

We still need to factor out things such as SSL key issuance, but we do
have the letsencrypt.nix module to do that, which we're going to use
very soon.

Another thing that's missing is a module which simulates a recursive DNS
resolver across the whole test network.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>