tests/headcounter/escalus: Register a few users
We want to do blackbox testing of the XMPP deployment, so we need some
pre-existing users.

Escalus has functionality to do that but it requires direct access to
the running MongooseIM instance (either locally or via distributed
Erlang), so we're running a series of mongooseimctl calls on the host to
ensure Escalus has *no* means to access the node at all.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>