Add a module for service startup conditions
The module generates services that order themselves before the systemd
service of the same name and run a small C program which checks for
whether an address is bindable or a TCP port is reachable.

This avoids race conditions during startup, because services can now
reliably depend on other services that do not support socket activation.

In production we have a dependency on keys.target on most of these
services, so it's not a very big issue there (even though it's definitly
nicer to avoid this issue by using this module), however on tests this
is cruicial, because we currently don't have a service that feeds keys.

Even if we'd have that service, it would need to be synchronous across
all machines, still leaving the same issue in place.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>