ultron: Create Hydra database on PostgreSQL init
It's been a long time since we ran PostgreSQL on taalo and it has been
moved to ultron since cea50f1e7583e2908313ea00c5d707136ab42226.

While doing so in production I just copied the database over to ultron,
however if we'd want to set up the whole network from scratch it would
fail (and it does so in our new tests) because hydra-init expects the
database to already exist.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>