tests/headcounter: Increase service timeouts
Some tests are failing when Hydra runs all of the tests in parallel
because the VMs generate a pretty high load on benteflork.

This works around the issue by setting the default timeouts to 600
seconds and overrides the PostgreSQL "TimeoutSec" value as well.

The reason for the PostgreSQL exception is that in the NixOS module the
TimeoutSec value is set so the defaults get overridden in that specific
case.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>