tests/headcounter: Don't change DefaultRestartSec
A very bad idea to change this, because the SSH tunnels depend on retry
and the default value is 200ms for a reason. So we end up making things
much worse, because our conditions checkers *and* all the other services
have now the same timeout so a single Tunnel connection failure will
block all of our tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>