common.nix: Add kernel patch fixing tun poll()
The patch is currently in testing and will be submitted by Hannes
Frederic Sowa to the netdev mailing list, so this patch is preliminary.

I've added an assert on NixOS release 16.09, so that this will
deliberately break eval as soon as we switch to 17.03 because first of
all we will have boot.kernelPatches then and it's also very likely that
this patch isn't needed anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>