tests/headcounter: Exclude certain nodes
We really don't need taalo, benteflork and unzervalt to do XMPP tests
and it only increases load and startup time during the tests.

So we now have a excludeNodes attribute for Headcounter specific tests
that cause these nodes to be built but not started up by default, pretty
much similar to the --exclude argument in nixops deploy.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>