modules/testing: Split off resolver from CA module
The resolver part makes more sense to be generally included in VM tests,
so that every test node has a real simulation of Internet DNS root
servers.

So far we have plunged a bunch of FQDNs into the hosts file of a few
specific nodes, but having the whole DNS infrastructure available also
means that we can test our DNS setup properly as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>