services/dyndns: Factor out NSD zone writer
The zone writer command is going to be used for our upcoming ACME
implementation as well, so it makes sense to re-use the same code
instead of duplicating the implementation all over the place.

So now we have a generic zone writer, which is similar to what BIND has
with nsupdate.

Tested successfully against the dyndns NixOS test.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>