nsd-zone-writer: Properly isolate user permissions
We now use a dedicated user just for writing dynamic zones and we allow
only a list of users specified with allowedUsers to update zones.

The implementation is a bit more involved, because we only allow a
setuid binary to run as the dynzone user and because we can't use the
setuid bit on shell scripts, we use a small wrapper around that.

One of the reasons I'm not using the security.wrappers option but
implement this on my own is that we want our wrapper to be as simple as
possible without propagating *any* environment variables and without any
magic around figuring out the target binary from a common path.

I'm also using more strict permissions on everything related to the zone
writer, so that we only have those permissions set that are absolutely
necessary and nothing more.

To make sure this is the case we now have an extra test exclusively for
the NSD zone file writer.

Tested with the new nsd-zone-writer and the dyndns NixOS VM tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>