nsd-zone-writer: Remove NSD client snakeoil key
While NSD only accepts commands from the local system, in theory it
shouldn't be an issue havivng even a snakeoil key.

However, we're going to have another service (ACME) be able to write NSD
zones and we have properly separated the NSD zone writer from the rest
of the system so that only certain users may write zone files.

So with this in mind it really doesn't make sense if we have all that
local privilege separation while the NSD client keys are being
world-readable and in /nix/store.

This now generates the keys so that they're only readable by nsd and the
dynzone user, leading to proper separation even in the event that we got
a local compromise of the name servers.

Tested successfully with the nsd-zone-writer and the dyndns NixOS VM
tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>