dyndns: Factor out single-source Haskell compiler
We're going to need this for ACME as well, so similar to compileC we now
have compileHaskell, which allows us to either directly inline Haskell
code or reference a path.

In both cases it spins up a GHC with the right buildDepends and compiles
the file specified by the source attribute.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>