pkgs: Add a small library for systemd services
This library is basically a set of helpers to build services that nicely
integrate into systemd's socket activation.

I found myself just about to start another service which by its base
would be more or less a copy of dyndns.hs, but instead of copy & pasting
stuff around, I thought it would be better to actually write something
that could serve as a better base than network-simple.

So there is the "daemons" package, which could help but unfortunately it
still doesn't provide socket activation, nor does it (allow to) set
SO_BINDTODEVICE for client sockets.

And regarding socket activation itself, I haven't found a lot more on
Hackage except the most basic bindings. Even the "systemd" package that
I'm using had to be bumped to 1.1.2 (because we're not on 17.03 yet) in
order to allow named sockets.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>