compile-haskell: Allow to add drvs as buildDepends
So far we're only using strings, which are used as attribute keys to
look up on haskellPackages. But in order to be able to use our shiny new
nexus package, we either would need to override haskellPackages and
inject nexus there or provide a way to pass it directly to
compileHaskell.

I choose the latter one, because it is simpler.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>