services/dyndns: Refactor to use socket activation
We now use our new nexus library to re-use the socket file descriptors
passed to us via systemd.

This gets rid of the port/host options in the YAML config file and moves
this solely to the service module.

However, we still have a problem here because the master needs to wait
for _all_ slaves to become available, which in production is a bit error
prone, because the sockets for the master need to be delayed until the
tunnels are up.

So if one of the slaves fail to start, the master won't come up. This is
the opposite of robustness, so we need the master to connect to all of
the slaves, and if a failing slave becomes available again we're sending
sync packets anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>