pkgs/nexus: Acknowledge sent packages
In the dyndns implementation, we're only sending data from the master
process but never receive anything. In theory this is fine, but in
practice this means that we won't notice disconnects if the slave
process closes the connection.

So after every packet received, we're now sending an '@' character which
the send function checks directly after sending.

This also means that we need to fix our tests, not only because send now
returns Bool but also because all sends are now synchronous and don't
return immediately anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>