services/dyndns: Refactor to use slave listeners
So far we had listeners for slave connections on the master, but that
turned out to be a little bit not so practical because this means that
all of the configured slaves need to be available in order for the
master to function correctly.

Now this is the opposite and we have a quite nice implementation that's
entirely race-free and socket-actived based on whether the specific
slave devices are available or not.

We're also using SO_BINDTODEVICE whenever we have a specific device set,
which is quite crucial for our encrypted tunnels if for example we have
an externally routable address that collides with one of our tunnel
addresses. The latter would essentially mean that we'd send out our
stuff via an unencrypted connection, which of course we don't want to
happen at all. Using SO_BINDTODEVICE prevents this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>