nexus: Implement preliminary zone file types
Right now this is just a bunch of data types that we're going to need in
order to write zone files. The DomainName parser right now is a bit
limited because it doesn't allow wildcard records or underscores
typically used in SRV records.

This also includes a safecopy migration instance for the old version of
Zone we currently have in dyndns.hs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>