nexus: Factor out types into own namespace
So far we had some custom types spread around in the module tree and
another set of custom types in DNS.hs.

From the point of just using this library it makes no sense to expose
the whole bunch of internal stuff, so we're only re-exporting the
important parts from DNS.hs and leave all the rest in Types.hs and in
Types/*.hs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>