nexus: Switch to using lenses
Traversing all these nested data structures is a pain in the ass, so
we're going to use lenses to make modifying them *way* easier.

So far we're not using them extensively except for updateRecords and
even the latter isn't using it to its full extent (for example we can
match the Nameserver by using Prism).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>