nexus: Add preliminary zone renderer
Represents the main functionality of what we want to achieve with our
Zone data type, because what we actually want is to generate zone files
out of a set of (nested) data types.

This is implemented as a ByteString builder, but it currently lacks
escaping for stuff like TXT records.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>