nexus: Factor out Domain type into a type class
The Domain sum type is really more of an obstactle rather than being
useful.

We now have a RRName type which is useful for every domain specification
where special zone file syntax is allowed, like for example wildcard
records or "@" to reference the origin.

Having this in a type class also allows us to easily extend the type.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>