nexus: Refactor and expose different Zone creators
This introduces a few helpers to create zones along with different
variations on the SOA records, so that we now can create such zones for
dyndns and for ACME with just very short function calls.

I've also renamed the fromFQDN function to toRRName, because the former
really is useless because it only converts to the type of the class in
use.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>