nexus: Make TextRecord a ByteString and encode
I was initially using Data.Text because it would fit to TXT records by
its name, but it's actually not what we want, because TXT records only
support ASCII characters (and escapes).

The escaping is now a little bit more correct, although it still has a
few issues:

  * Decimals need to be encoded with a fixed 3-digits padding according
    to RFC1035[1].
  * We need to break up strings with a length of > 255 characters into
    chunks.

For our use-case however, it's not a priority to resolve these issues.

[1]: https://tools.ietf.org/html/rfc1035#page-35

Signed-off-by: aszlig <aszlig@redmoonstudios.org>