nsd-zone-writer: Add ability to remove a zone
This is needed mainly for ACME in order to remove the zone after we put
out the challenge TXT record.

Even though nsd-control (according to its manpage) is not removing zones
that are configured using nsd.conf I'm making extra sure that the file
actually exists in the configured zone directory (usually called
"dynzones"), because we don't want to accidentally delete important
zone files.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>