nsd-zone-writer: Validate FQDN argument
The zone writer is a setgid program that's only executable by users in
the dynzone group and the only programs that currently have access to
that are the dyndns implementation and the upcoming ACME service.

While this is a relatively small attack surface, I think it's a good
idea to check the FQDN anyway in case one of these services make it
possible for attackers to inject FQDNs or if I somehow drive nuts and
make this possible for everyone to enter their custom zones.

So we're now checking the FQDN for valid FQDN characters and fail if
that's not the case.

In the event that an attacker could inject arbitrary characters here,
we'd be prone to a directory traversal vulnerability, because it's
possible to use .. to write the zone to any directory where the dynzone
user has access to. Currently it's just /var/lib/nsd/dynzones, but
better be safe than sorry.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>