dyndns: Factor out socket generators into hclib
Since we now have our standardized nexus library, it doesn't make sense
to replicate the exact same code that's providing socket units for a
nexus service all over the code base.

Tested via the dyndns NixOS VM test and it didn't need to be run because
the hash was the same, which implies that this commit doesn't change
anything in functionality.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>