pkgs: Add acmetool at version 0.0.59
I've decided to go with acmetool instead of the other options out there,
because most options are either written in a dynamically typed language
or don't support dns-01 validation or are doing more unnecessary things
we don't want.

So because we only need dns-01 validation, so this comes with a patch
which removes everything except the DNS responder from the acmetool code
base.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>