Fix main site application
This has been using a rather old version of Haxe. Unfortunately, newer
versions of Haxe do not work out of the box, so we need to fix a bunch
of things.

First of all, we need to use a newer version of Hase, because that
already contains a lot of fixes for newer Haxe versions and we're now
also removing the C++ target, because it was unused anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>