pkgs/acmetool: Add patch to set "desired" path
Unfortunately, acmetool does not allow to customize any paths that are
defined by their storage specification[1], except for the hooks
directory.

I guess reason the hooks directory can be changed is that it isn't based
on their "fdb" database-like interface and thus it's easy to implement.

So this patch lifts the constraints of the fdb interface by allowing
absolute paths to be used as subdirectory names.

This is very ugly, but these directories are hardcoded anyway, so for
our case it's fine. Unfortunately it's something that's unacceptable for
upstream, so we might want do a proper patch in the long term.

In summary: We now have an ACME_DESIRED_DIR environment variable that we
can set to point the "desired" directory to a store path.

[1]: https://github.com/hlandau/acme/blob/55dbf10a1a51bc11ccd1573efc38be6c3ce49c11/_doc/SCHEMA.md

Signed-off-by: aszlig <aszlig@redmoonstudios.org>