pkgs/acmetool: Add patch for absolute symlinks
The reason acmetool only allows relative symlinks is that it wants the
state directory to be portable so it can be moved around in the file
system hierarchy.

In our case, we want to use a bind mount for delaying services, but that
won't work if we have relative symlinks, except if we'd bind-mount the
whole state directory.

We don't want that, so this patch makes every symlink absolute.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>