Add acme service module and test
The service is very different to what we have in <nixpkgs> and the main
difference is that we're only focusing on dns-01 validation.

Another difference is that we don't need to take core of ordering,
because the way we expose certificates to services is using a bind mount
to /run/ssl-certs, which is delayed until the acme service has finished
getting at least the initial certificate.

Of course, right now we need to come up with an elegant solution on how
to do renewal without needing to add service reload definitions, but
that's something for another commit.

With our new nexus library in place, the main validation hook is using
it in order to communicate with the name server(s) and write the
generated zone files.

By design we expect all the _acme-challenge.x.y domains to reside in
their own zone, so it's easy to add and remove the zone without handling
too much state.

We're also using POSIX ACLs to give permissions to the certificate and
private key files, which might not be supported by all file systems.
However, I think this is less of an issue, because in the event a file
system is lacking support one can still point the stateDir to another
file system.

The VM test is using the webserver that's built into twisted, because I
haven't found a (packaged) webserver in <nixpkgs> which supports SSL and
without a whole bunch of configuration blurb. All of the web servers
currently having a NixOS service module don't actually set User/Group in
serviceConfig and they drop the privileges themselves (_after_ reading
the cert/key), so we can't really test whether the certificates got the
right permissions.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>