acme: Add options for reload/restart of services
While it is certainly possible to get out information about PIDs using
the autofs ioctl, it won't help us because we don't know whether to
reload or restart a particular service.

That's why I'm now introducing two options called "reloads" and
"restarts", which lists the units that need to be reloaded/restarted as
soon as there are changes in the ACME certificates.

The implementation is based on inotify and path units, which are
triggered as soon as there are changes on the certificate file. Those
units are also ordered to run _after_ the acme-setperms service, so that
no race conditions appear.

I've also added a test for this, so we're forwarding time closely
towards certificate expiry and checking whether the new certificate has
been issued and also whether the service has been restarted.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>