acme: Add handler option to use a specific device
In production we want to make sure that we *only* use an encrypted
tunnel device to connect to the validation handler, so let's add an
option for specifying a device similar to how it's already the case in
our dyndns implementation.

The test is now using a tunnel device (although unencrypted), which
should replicate what we want in our production setup.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>