testing/resolver: Filter out subdomains
We only want to have all the domains from all the nodes, but not
subdomains because we actually want to make sure delegation works.

This is especially cruicial for testing torservers.net, where we don't
have DNS access to the main domain and need to rely on delegation.

So within our test environment we want to make sure that we can't change
anything on torservers.net but only work on eg. jabber.torservers.net.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>