modules/testing: Add a tshark sniffer module
While debugging VM tests I frequently need to run a sniffer, which I
mostly did using something like this:

$machine->execute('tshark -n -i eth1 >&2 &');

Unfortunately this also includes adding tshark to systemPackages of the
host and I also need to manually specify the devices I want to sniff on
(yes, I know there is "any").

So now it's simply a matter of adding this module to the imports
attribute and we get sniffer output on all interfaces.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>