acme: Add hook event for adding all in one cert
We now have a "full" certificate which is comprised of the fullchain and
the private key. This is needed for things such as lighttpd which
doesn't have an extra option to specify the private key.

The VM test now also uses lighttpd to test for exactly that and also
uses otherDomains, so we have that covered now as well.

We now also check whether a user who shouldn't get access to private
keys (this includes "full") is able to access them.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>