tests/headcounter: Use virtual Internet DNS/IPs
So far our tests were a bit limited and we have only used ultron's
vhosts (but not its main IP) in all of the test networks nodes by
creating static routes to it.

However, this worked only by accident because ultron still was reachable
via the primaryIPAddress created by the NixOS test system.

Now we're giving all machines of the deployment the same address that
they have in production (although with a different prefix length) and
automatically assign addresses to non-deployment nodes.

For the automatic assignment, I'm not using 192.168.1.x but 93.184.216.x
(example.com has 93.184.216.34, so that's why I'm using this, because
it's not very likely that we'll ever get assigned a public address like
this) and fdbc::x (just random) instead, because it's easier to spot
whether something went wrong and we still have 192.168.1.x addresses
somewhere showing up.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>