Switch deployment and tests to use ACME
This is the final step for switching the deployment to use ACME in
conjuction with DNS activation.

It still has some rough edges and we still need to refactor a bit before
we finally push this to production, but we now not only have an
implementation but also a very realistic test system for running
production tests with ACME certificates (which are also in use for our
xmppoke tests).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>