tests/xmppoke: Check availability of SRV host
So far we've used a simple ping/netcat on the FQDN directly, but for
torservers.net this won't work, because we're not in control of the main
domain.

The reason why this has worked in the past was that we have iterated
through all of the FQDNs (where torservers.net is included, because it's
actually a virtual host) and built up a static host entry for all of the
nodes.

With real DNS in place, we also have a much more realistic test in
place, which also means that the FQDN we're testing is no longer the
real XMPP server.

Now we're carving out the real FQDN of the XMPP server using an SRV
lookup.

Unfortunately, the test still doesn't pass because there still seems to
be a resolving issue with xmppoke as it tries to connect to
torservers.net directly instead of jabber.torservers.net.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>