tests/xmppoke: Use nest instead of subtest
A minor annoyance of subtest is that we get an output like this
regardless of whether the whole test failed or not:

1 out of 1 tests succeeded

The reason for this is that we try to assure that the test derivation
will always succeed to build, so we still get a nice test report if any
of these tests have failed.

So let's just nest one level deeper instead of creating a subtest.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>