tests: Simulate torservers.net more accurately
We now have a webserver in our test network which resembles the ACME
setup of torservers.net. The validation of the certificates is done
using the ACME http-01 validation method and we *might* be able to use
dns-01 validation independent of that.

To be absolutely sure, we need to update boulder to the latest upstream
production release, but so far this seemed to work even though we have a
different account key for the same domain.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>